Added the capacity for enemies to damage the player through either projectiles, or collisions.

+++ sporebornGame/Assets/Scripts/Models/HealthModel.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class HealthModel : MonoBehaviour
 {
-    public int maxHealth;
-    public int currHealth;
+    public float maxHealth;
+    public float currHealth;
 
     public SpriteRenderer spriteRenderer;
